refactor: extract JiraIssueExportTarget.swift from IssueExportTargets.swift
Splits the `JiraIssueExportTarget` Codable target out of
IssueExportTargets.swift, mirroring #815 GitHubIssueExportTarget.

Byte-identical move of lines 32-64. Both files import Foundation.

IssueExportTargets.swift: 124 → 90 lines. Tests: 667.
